[WIP] Support tensors as Dict keys
This prepares the PR where we implement sets in terms of dicts.

The Hashable variable is opaque on purpose so that it fails hard if
if it unadvertedly leaks back into user code.

We also found and fixed a number of latent bugs and inconsistencies
in the way dynamo checked what can be a dict key.

Fixes https://github.com/pytorch/pytorch/issues/107595
Fixes https://github.com/pytorch/pytorch/issues/111603

Fix import

ghstack-source-id: d83d34ca90976179fb0a46711333f1e8f497f40c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116672

rebase, fixes