[export] do not unlift cond/map submodules
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118610

I don't think we should be unlifting HOO submodules.

What is the constract of unlifting? It is: restore the original calling convention of the module, undoing the transformation in which we lift parameters, buffers, and constants to inputs in the graph.

Unlifting does *not* make any guarantees about what's going on inside the module. It's still a flat module. So why should we lift the cond/map submodules? It doesn't have anything to do with the contract stated above; it's some internal stuff that doesn't affect how the module will be called.

Further, this code as written modifies the state dict; adding a new buffer that is actually duplicate of a previous buffer. Modifying the state dict from the original eager module is never correct.
ghstack-source-id: 213477244
@exported-using-ghexport

Differential Revision: [D53160713](https://our.internmc.facebook.com/intern/diff/D53160713/)