[export] do not rewrite state dict when unlifting
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118611

This is Very Bad; changing state dict keys violates one of the key contracts we have, which is "do not mess with the state dict".

Change unlift to use a similar `_assign_attr` approach that fx.GraphModule and unflatten do.

Also took the opportunity to improve the interface of `_assign_attr` to be more general.
ghstack-source-id: 213477245
@exported-using-ghexport

Differential Revision: [D53139277](https://our.internmc.facebook.com/intern/diff/D53139277/)