[inductor] more accurate throughput calculations for kernel benchmarks
Our current throughput calculations for kernel benchmarks have some issues,
particularly when we slice inputs in the kernel. In such cases, we count
the original inputs as part of the memory traffic passed across the kernel.
This is incorrect because it may result in a much larger throughput
calculation, which can even exceed the theoretical bandwidth.

Instead, we should only count the size of the "slices" that contribute to
the actual memory traffic.

ghstack-source-id: 50d806eff7f098b77700282a895b3d628b0a88f9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118858